Adjust logic in cor.h, corhdr.h, corinfo.h, corjit.h and such so that they may be included without the PAL (#46055)
This was done via simple substitution of the standard sized integer types, and dealing with the fallout. It is tested by using the ICorJitInfo interface directly within the crossgen2 jitinterface thunk library.

Effort was made to use a minimum number of casts, as casts in such a large change are likely to be wrong somewhere.
- Exceptions are the use of casting to handle some calls to the existing internal metadata api and around use of char16_t.

In addition, a small amount of logic from the PAL headers were pulled into the proper header to allow compilation in the presence of some SAL annotation, and other small details.

+++ src/coreclr/ToolBox/superpmi/superpmi/jitinstance.cpp
@@ -303,8 +303,8 @@ JitInstance::Result JitInstance::CompileMethod(MethodContext* MethodToCompile, i
 
     PAL_TRY(Param*, pParam, &param)
     {
-        BYTE* NEntryBlock    = nullptr;
-        ULONG NCodeSizeBlock = 0;
+        uint8_t* NEntryBlock    = nullptr;
+        uint32_t NCodeSizeBlock = 0;
 
         pParam->pThis->mc->repCompileMethod(&pParam->info, &pParam->flags);
         if (pParam->collectThroughput)
@@ -381,8 +381,8 @@ JitInstance::Result JitInstance::CompileMethod(MethodContext* MethodToCompile, i
 
 void JitInstance::timeResult(CORINFO_METHOD_INFO info, unsigned flags)
 {
-    BYTE* NEntryBlock    = nullptr;
-    ULONG NCodeSizeBlock = 0;
+    uint8_t* NEntryBlock    = nullptr;
+    uint32_t NCodeSizeBlock = 0;
 
     int sampleSize = 10;
     // Save 2 smallest times. To help reduce noise, we will look at the closest pair of these.
